use intptr_t for storing the integers in the void*-based StringMap; hopefully gets around inappropriate use of long on Windows
git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/rsbml@94668 bc3139a8-67e5-0310-9ffc-ced21a209358

+++ src/graph.c
@@ -1,6 +1,8 @@
 #include "rsbml.h"
 #include "StringMap.h"
 
+#include <stdint.h>
+
 #ifdef LIBSBML3
 #define SimpleSpeciesReference_getSpecies(x) SpeciesReference_getSpecies(x)
 typedef SpeciesReference_t SimpleSpeciesReference_t;
@@ -62,18 +64,18 @@ rsbml_build_graph(SBMLDocument_t *doc)
     int num_modifiers = Reaction_getNumModifiers(reaction);
     const char *id = Reaction_isSetId(reaction) ? Reaction_getId(reaction) : 
       Reaction_getName(reaction); 
-    long reaction_index = (long) StringMap_get(id_map, id);
+    intptr_t reaction_index = (intptr_t) StringMap_get(id_map, id);
     out_count[reaction_index] = Reaction_getNumProducts(reaction);
     for (j = 0; j < num_reactants; j++) {
       const char *id = SimpleSpeciesReference_getSpecies(
         (SimpleSpeciesReference_t *)Reaction_getReactant(reaction, j));
-      long index = (long) StringMap_get(id_map, id);
+      intptr_t index = (intptr_t) StringMap_get(id_map, id);
       out_count[index]++;
     }
     for (j = 0; j < num_modifiers; j++) {
       const char *id = SimpleSpeciesReference_getSpecies(
         (SimpleSpeciesReference_t *)Reaction_getModifier(reaction, j));
-      long index = (long) StringMap_get(id_map, id);
+      intptr_t index = (intptr_t) StringMap_get(id_map, id);
       out_count[index]++;
     }
   }
@@ -96,26 +98,26 @@ rsbml_build_graph(SBMLDocument_t *doc)
     int num_products = Reaction_getNumProducts(reaction);
     const char *id = Reaction_isSetId(reaction) ? Reaction_getId(reaction) : 
       Reaction_getName(reaction); 
-    long reaction_index = (long) StringMap_get(id_map, id);
+    intptr_t reaction_index = (intptr_t) StringMap_get(id_map, id);
     SEXP reaction_out = VECTOR_ELT(VECTOR_ELT(r_edges, reaction_index), 0);
     for (j = 0; j < num_reactants; j++) {
       const char *id = SimpleSpeciesReference_getSpecies(
         (SimpleSpeciesReference_t *)Reaction_getReactant(reaction, j));
-      long index = (long) StringMap_get(id_map, id);
+      intptr_t index = (intptr_t) StringMap_get(id_map, id);
       //Rprintf("reactant index: %d edge count: %d\n", index, out_count[index]);
       INTEGER(VECTOR_ELT(VECTOR_ELT(r_edges, index), 0))[out_pos[index]++] = reaction_index + 1;
     }
     for (j = 0; j < num_modifiers; j++) {
       const char *id = SimpleSpeciesReference_getSpecies(
         (SimpleSpeciesReference_t *)Reaction_getModifier(reaction, j));
-      long index = (long) StringMap_get(id_map, id);
+      intptr_t index = (intptr_t) StringMap_get(id_map, id);
       //Rprintf("modifier index: %d edge count: %d\n", index, out_count[index]);
       INTEGER(VECTOR_ELT(VECTOR_ELT(r_edges, index), 0))[out_pos[index]++] = reaction_index + 1;
     }
     for (j = 0; j < num_products; j++) {
       const char *id = SimpleSpeciesReference_getSpecies(
         (SimpleSpeciesReference_t *)Reaction_getProduct(reaction, j));
-      long index = (long) StringMap_get(id_map, id);
+      intptr_t index = (intptr_t) StringMap_get(id_map, id);
       //Rprintf("product index: %d\n", index);
       INTEGER(reaction_out)[j] = index + 1;
     }
